gists: Move `list_user_gists` to `GistsHandler`
GitHub keeps this endpoint with other gist related endpoints so we will
do the same for minimizing surprise.
