+++ src/api.rs
@@ -16,5 +16,4 @@ pub mod ratelimit;
 pub mod repos;
 pub mod search;
 pub mod teams;
-pub mod users;
 pub mod workflows;
